refactor: add error handling and input validation to auth config
