+++ src/config/frontendConfig.ts
@@ -1,14 +1,4 @@
-import { Roles } from '@/constants/roles';
 import { STORES, dbSet } from '@/lib/indexeddb';
-import { createProfile, getProfileByIdentifier } from '@/services/profile';
-import { mergeNames } from '@/utils/helper';
-import { extractSafeRedirectPath } from '@/utils/redirect-guard';
-import {
-  clearRedirectIntent,
-  getIntent,
-  getRedirectIntent
-} from '@/utils/redirect-intent';
-import { Patient, Practitioner } from 'fhir/r4';
 import { useRouter } from 'next/navigation';
 import React from 'react';
 import { SuperTokensConfig } from 'supertokens-auth-react/lib/build/types';
@@ -18,6 +8,11 @@ import ThirdParty from 'supertokens-auth-react/recipe/thirdparty';
 import { getClaimValue } from 'supertokens-web-js/recipe/session';
 import { UserRoleClaim } from 'supertokens-web-js/recipe/userroles';
 import { getAppInfo } from './appInfo';
+import {
+  handleNewUserLogin,
+  handleReturningUserLogin,
+  resolvePostLoginRedirect
+} from './auth-helpers';
 
 const routerInfo: { router?: ReturnType<typeof useRouter>; pathName?: string } =
   {};
@@ -28,136 +23,6 @@ export function setRouter(
 ) {
   routerInfo.router = router;
   routerInfo.pathName = pathName;
-}
-
-/** Fetches a CSRF token from the server for use in POST /auth/cookie requests. */
-async function fetchCSRFToken(): Promise<string | null> {
-  try {
-    const res = await fetch('/auth/cookie/csrf-token');
-    if (!res.ok) return null;
-    const data = (await res.json()) as { token?: string };
-    return data.token ?? null;
-  } catch {
-    return null;
-  }
-}
-
-/** Posts auth cookie data to the server with CSRF protection. */
-async function postAuthCookie(
-  body: Record<string, unknown>
-): Promise<Response> {
-  const token = await fetchCSRFToken();
-  const headers: Record<string, string> = {
-    'Content-Type': 'application/json'
-  };
-  if (token) headers['X-CSRF-Token'] = token;
-  return fetch('/auth/cookie', {
-    method: 'POST',
-    headers,
-    body: JSON.stringify(body)
-  });
-}
-
-async function postAuthCookieForUser(
-  role: string,
-  userId: string,
-  roles: string[] | undefined,
-  emails: string[],
-  phoneNumbers: string[],
-  profile: Patient | Practitioner | null
-): Promise<void> {
-  const cookieData = {
-    userId,
-    roles,
-    role_name: role,
-    email: emails[0] || '',
-    phoneNumber: phoneNumbers[0] || '',
-    profile_picture: profile?.photo?.[0]?.url ?? '',
-    fullname: mergeNames(profile?.name),
-    fhirId: profile?.id ?? ''
-  };
-  const cookieRes = await postAuthCookie(cookieData);
-  if (!cookieRes.ok)
-    console.error('[auth:cookie] server returned', cookieRes.status);
-}
-
-/** Handles login for new users — creates FHIR profile if missing, sets auth cookie. */
-async function handleNewUserLogin(
-  roles: string[] | undefined,
-  userId: string,
-  emails: string[],
-  phoneNumbers: string[]
-): Promise<void> {
-  const role =
-    Array.isArray(roles) && roles.includes(Roles.Practitioner)
-      ? Roles.Practitioner
-      : Roles.Patient;
-  let profileData: Patient | Practitioner | null = await getProfileByIdentifier(
-    { userId, type: role }
-  );
-
-  if (!profileData) {
-    await createProfile({
-      userId,
-      email: emails[0] || '',
-      phoneNumber: phoneNumbers[0] || '',
-      type: role
-    });
-    profileData = await getProfileByIdentifier({ userId, type: role });
-    if (!profileData) throw new Error('Failed to create profile');
-  }
-
-  await postAuthCookieForUser(
-    role,
-    userId,
-    roles,
-    emails,
-    phoneNumbers,
-    profileData
-  );
-}
-
-/** Handles login for returning users — fetches FHIR profile and sets auth cookie. */
-async function handleReturningUserLogin(
-  roles: string[] | undefined,
-  userId: string,
-  emails: string[],
-  phoneNumbers: string[]
-): Promise<void> {
-  const role =
-    Array.isArray(roles) && roles.includes(Roles.Practitioner)
-      ? Roles.Practitioner
-      : Roles.Patient;
-  const profile: Patient | Practitioner | null = await getProfileByIdentifier({
-    userId,
-    type: role
-  });
-
-  await postAuthCookieForUser(
-    role,
-    userId,
-    roles,
-    emails,
-    phoneNumbers,
-    profile
-  );
-}
-
-/** Resolves post-login redirect URL from stored intent or query params. */
-function resolvePostLoginRedirect(): string | null {
-  const redirectUrl = getRedirectIntent();
-  if (redirectUrl) {
-    clearRedirectIntent();
-    return extractSafeRedirectPath(
-      `?redirectToPath=${encodeURIComponent(redirectUrl)}`
-    );
-  }
-  const intent = getIntent();
-  if (intent) {
-    clearRedirectIntent();
-    return (intent.payload?.path as string) ?? '/';
-  }
-  return extractSafeRedirectPath(globalThis.location.search);
 }
 
 export const frontendConfig = (): SuperTokensConfig => {
