+++ web/auth-spa/src/frontendConfig.tsx
@@ -1,146 +1,14 @@
-import type { Patient, Practitioner } from 'fhir/r4';
 import type { SuperTokensConfig } from 'supertokens-auth-react/lib/build/types';
 import Passwordless from 'supertokens-auth-react/recipe/passwordless';
 import Session, { getClaimValue } from 'supertokens-auth-react/recipe/session';
 import ThirdParty from 'supertokens-auth-react/recipe/thirdparty';
 import { UserRoleClaim } from 'supertokens-auth-react/recipe/userroles';
 import { getAppInfo } from './appInfo';
-import { Roles } from './constants/roles';
-import { createProfile, getProfileByIdentifier } from './services/profile';
-import { mergeNames } from './utils/helper';
-import { extractSafeRedirectPath } from './utils/redirect-guard';
 import {
-  clearRedirectIntent,
-  getIntent,
-  getRedirectIntent
-} from './utils/redirect-intent';
-
-/** Posts auth cookie data to the server with CSRF protection. */
-async function postAuthCookieForUser(
-  role: string,
-  userId: string,
-  roles: string[] | undefined,
-  emails: string[],
-  phoneNumbers: string[],
-  profile: Patient | Practitioner | null
-): Promise<void> {
-  const cookieData = {
-    userId,
-    roles,
-    role_name: role,
-    email: emails[0] || '',
-    phoneNumber: phoneNumbers[0] || '',
-    profile_picture: profile?.photo?.[0]?.url ?? '',
-    fullname: mergeNames(profile?.name),
-    fhirId: profile?.id ?? ''
-  };
-  const cookieRes = await postAuthCookie(cookieData);
-  if (!cookieRes.ok)
-    console.error('[auth:cookie] server returned', cookieRes.status);
-}
-
-async function handleNewUserLogin(
-  roles: string[] | undefined,
-  userId: string,
-  emails: string[],
-  phoneNumbers: string[]
-): Promise<void> {
-  const role =
-    Array.isArray(roles) && roles.includes(Roles.Practitioner)
-      ? Roles.Practitioner
-      : Roles.Patient;
-  let profileData: Patient | Practitioner | null = await getProfileByIdentifier(
-    { userId, type: role }
-  );
-
-  if (!profileData) {
-    await createProfile({
-      userId,
-      email: emails[0] || '',
-      phoneNumber: phoneNumbers[0] || '',
-      type: role
-    });
-    profileData = await getProfileByIdentifier({ userId, type: role });
-    if (!profileData) throw new Error('Failed to create profile');
-  }
-
-  await postAuthCookieForUser(
-    role,
-    userId,
-    roles,
-    emails,
-    phoneNumbers,
-    profileData
-  );
-}
-
-/** Handles login for returning users — fetches FHIR profile and sets auth cookie. */
-async function handleReturningUserLogin(
-  roles: string[] | undefined,
-  userId: string,
-  emails: string[],
-  phoneNumbers: string[]
-): Promise<void> {
-  const role =
-    Array.isArray(roles) && roles.includes(Roles.Practitioner)
-      ? Roles.Practitioner
-      : Roles.Patient;
-  const profile: Patient | Practitioner | null = await getProfileByIdentifier({
-    userId,
-    type: role
-  });
-
-  await postAuthCookieForUser(
-    role,
-    userId,
-    roles,
-    emails,
-    phoneNumbers,
-    profile
-  );
-}
-
-/** Resolves post-login redirect URL from stored intent or query params. */
-function resolvePostLoginRedirect(): string | null {
-  const redirectUrl = getRedirectIntent();
-  if (redirectUrl) {
-    clearRedirectIntent();
-    return extractSafeRedirectPath(
-      `?redirectToPath=${encodeURIComponent(redirectUrl)}`
-    );
-  }
-  const intent = getIntent();
-  if (intent) {
-    clearRedirectIntent();
-    return (intent.payload?.path as string) ?? '/';
-  }
-  return extractSafeRedirectPath(globalThis.location.search);
-}
-
-/** Posts auth cookie data to the server with CSRF protection. */
-async function postAuthCookie(
-  body: Record<string, unknown>
-): Promise<Response> {
-  let token = '';
-  try {
-    const res = await fetch('/auth/cookie/csrf-token');
-    if (res.ok) {
-      const data = (await res.json()) as { token?: string };
-      token = data.token ?? '';
-    }
-  } catch {
-    /* CSRF token fetch is best-effort */
-  }
-  const headers: Record<string, string> = {
-    'Content-Type': 'application/json'
-  };
-  if (token) headers['X-CSRF-Token'] = token;
-  return fetch('/auth/cookie', {
-    method: 'POST',
-    headers,
-    body: JSON.stringify(body)
-  });
-}
+  handleNewUserLogin,
+  handleReturningUserLogin,
+  resolvePostLoginRedirect
+} from './auth-helpers';
 
 /** SuperTokens frontend configuration. */
 export const frontendConfig = (): SuperTokensConfig => {
